:recycle: [#18] -- move error list into re-usable utility component
This allows for re-using the error list between wrapper and 'atomic'
components, especially for the multiple: true variant.

NOTE: form.io itself always displays the errors at the wrapper level,
but it should be better UX/a11y to display the errors precisely with
the invalid input. There is definitely more to gain in terms of a11y
here, e.g. with aria-describedby etc.