:boom: Remove multiple for number component
After discussion with our support engineer and checking our saas
environments for usage of number component with multiple=True, it's
clear nobody actually uses this feature. Since it complicates the
handling of null/undefined, we can avoid dealing with this by just
removing that functionality (if anyone misses it, we can then solve
it when there's a stakeholder for it).