:sparkles: [#31] Initialize field values for constraint validator
When the mode of a min/max date constraint validation configuration
changes, other fields become relevant. E.g. 'future' or 'past' can
specify if today should be included or not, while 'fixedValue' does
not require any such extra input.

What exactly must be configured for each mode is declared in a type-safe
way to ensure it's complete and as concise as possible.

The select dropdown now accepts an optional onChange callback so you
can run additional callbacks on change events, which is particularly
interesting with React's batching of state updates and reduces the
amount of re-renders and/or ambiguous transitive states.