:sparkles: Pretend that inputMask is supported
inputMask is used under the hood by the postcode component. The formio
builder implementation stops at the placeholder. Since these components
are not actually used in the renderer/SDK and are only set up for
preview purposes and it's *too much* work to get the masking hooked
up, I've decided to let it be.