:sparkles: [#52] Add formio textarea component
This was originally planned to use for the itemsExpression
in the selectboxes component, but turned out to not be easy
to manage to validate for valid JSON and eventually
replaced with the already existing (courtesy of Laurens)
JSONEdit component.

Since we need to eventually implement the TextArea builder
component anyway, it will be of good use later and was
decided to keep it.