:sparkles: [#52] Implement the preview of a selectboxes component
When manual values are available, they are rendered as is and
their default value is taken into account.

When the variable data source is specified, a single option
is rendered to at least inform the form builder of the
expression being used.

In the future, we can possibly provide a 'variable' box so
that they can try out the preview with a realistic form
variable state.