:ok_hand: [#72] PR feedback
* Remove total from preview to reduce risk of confusion
* Fix validation error being mangled for array field.

The root cause of the valiation error mangling is a bug/shortcoming in
Formik's FieldArray which (too eagerly) splits the validation error
string into an array of characters and updates that based on the index
of an item being removed. See https://github.com/jaredpalmer/formik/issues/3352
for an existing bug report.