:recycle: [#81] Move another flaky test over to jest
Note that the components fetch available dropdown options asynchronously, which leads to
act warnings typically because the state updates happen outside of the initial
render and are thus unexpected.

These useEffect-based things run despite there not being any explicit user interaction
(which is automatically wrapped in act via @testing-library/user-event). Instead, we
can set up the jest fake timers so that we can still assert the loading state before
it resolves, perform the user interactions and use the async queries to assert the
expected final state, without having any act warnings.

To be fair I don't 100% understand how the timer advancing works, but the tests
seem to work when putting assertions in that should not pass (the test fails
as expected) and we can assert the in-between/loading states, without getting
any act warnings, so I'll take it.