:green_heart: [#89] Fix jest tests
The bundling/export is messed up, so importing @open-formulieren/leaflet-tools does not
perform the expected side effects in a Jest context, causing cryptic crashes. It seems
fine in Storybook and actually bundled code.

For more information, see https://github.com/kartena/Proj4Leaflet/pull/147.

If I was a beer drinker, this would be a good excuse to get properly intoxicated.