:recycle: [#32] Refactor Paths utility type
The ExcludeBareObject based on Record<string, unknown> type
was not wide enough - it was ignoring well-defined interfaces.

This is a Typescript thing - while the interface is assignable
to the Object/record, it's not considered so by typescript
because an interface lacks the indexed access ([K: string]
or similar). This manifests in the Radio/Select/Selectboxes
schemas because they amend the openForms extension (and
make it required), changing the level where the interface
is picked up and not recursing into its keys.

This check was added in the past to prevent recursion
issues with Typescript due to Formio's bare Object use
in types, like . This appears to be
mitigated now by checking T[K] for being a primitive
and only recursing in the other case.