:sparkles: Implement a modal component with bootstrap styles
Formio's modal component is based on or inspired by bootstrap. To make
transitioning easier, we apply the same formio classnames so that we
can share the styling.

The modal implementation needs some tweaking to match the markup used
in Formio/bootstrap, but other than that it's pretty straight forward.

The ModalContext primarily serves to mute warnings and make interaction
testing easier in Storybook, by:

* disabling the aria-hidden integration on the app element
* injecting the modal node in the right place in the DOM so that it's
  within the canvas