:recycle: Extracted the component preview toggle
The preview toggle is re-implemented based on the toggle mode component
so that it can be re-used in the preview panel and the content edit
form.

The component is moved to its own component file to declutter and make
it more clear where each component is implemented.

The ComponentPreview is cleaned up to make use of the new component,
and the same applies for the Content edit form.