:alembic: Implement generic top-level components for form previews
The form preview is broken out in two possible render modes, each have
their own top-level component to lock in the context and behaviours.

The structure mode is the most complex, since it will integrate with
the edit modal/form for each individual component, and the controls
for position within the component tree.