:truck: [open-formulieren/open-forms#4546] Move RichText component to 'base building blocks'
While it initially was only relevant for the content component, we will
make use of this same component in the display component for the
soft required validation errors.

The component translations wrapper around it remains component-type
specific since we can leverage the type-safety when checking validation
error keys, at the cost of a little bit of code duplication.