:sparkles: [open-formulieren/open-forms#4546] Implement new formio component to display soft required validation errors
The component is similar to the content component, but needs to be its
own type so that we can process it in the SDK and find a target to
display the form validation errors.

The layout/warning scaffolding will also be fixed, and it has no
meaning in the backend data, so a number of configuration options from
the content component are not available (such as conditional logic,
hidden/visible, display configuration, custom class).