:rotating_light: [open-formulieren/open-forms#2177] Patching react-leaflet-draw
The React-leaflet-draw package seems abandoned. As such, a fix for the typescript issue (https://github.com/alex3165/react-leaflet-draw/issues/188) probably won't happen anytime soon

The .patch file, in `/patches`, contains the fix needed for the react-leaflet-draw index.d.ts file.

The problem with this solution, is that this only solves it for this project. If we introduce typescript to the SDK (or any other project using `react-leaflet-draw`), this problem will need to be solved again