:art: Address jest tests warnings
Some blur behaviour is leading to state updates in Formik (likely
due to validation being triggered), which wasn't caught in act.

I'm not sure why the  interaction isn't picking this
up, it may heave to do with timers again... Wrapping it in act
at least makes the warning go away.