:sparkles: [open-formulieren/open-forms#5575] Add `headerEnd` to Panel component
Using `headerEnd` we can add additional content (like texts, icons, buttons) to the Panel header. This additional content will be placed at the end of the header (so in an LTR this will be on the right side of the header).

To make sure that the collapse clickable area still takes a majority of the available Panel header space, while not interfering with whatever additional content was added, the clickable area has been moved to a different div element. Using the Bootstrap card styling, this new clickable area still covers most of the card-header space.