:test_tube: [#236] Added failing test for Select component bug
The order of Select items should be effected by when you select an option. When looking at the actual data, the selected values are indeed in the correct order.

The Select component uses `extractSelectedValue` to get the actual selected values from the available options. The problem is, that it uses the order of the options when extracting the values. Meaning that the order will always be the same, regardless of user actions/input.