:safety_vest: [open-formulieren/open-forms#5708] Fixing profile validation schema
The profile validation schema needs to allow `null` as this is used by formio as "empty" value for falsy `defaultValue` values.

To simplify the validation schema, the support for defaultValue is removed. This should realistically never happen.
(There are no defaultValue configuration fields, and defaultValue won't really be used due to the prefilling from the communication preferences prefill plugin)