:truck: [#5723] Update map overlay type
Renaming the map overlay type, and adding an internal ID.

Previously, the `url` of the overlay type was removed from the type definition. This was done, because the url is only required for the frontend. The backend dynamically sets the url when fetching the components for SDK usage.

Now, instead of hiding the `url` property, we simply set an empty string. The backend still overrides this when passing the component information to the SDK.

The `_OF_INTERNAL_id` will be used to resolve the bug of this ticket. `_OF_INTERNAL_id` will be a stable identifier, which will be used to ensure that the overlay components don't re-render with every title or tile layer change.