:bug: [#5723] Prevent overlay menu from automatically closing
The map overlay configuration menu would automatically close when the WMS layer or overlay title were changed. This automatic closing was a result of the `OverlayTileLayer` key changing, which re-mounted the component.

By adding an internal, form-builder-specific, identifier, we have a stable value which can be used as the component key. This ensures that the component doesn't re-mount on every title of tile layer change.