:children_crossing: [#5247] Dynamic map lat/lng field steps
By default, number fields use a `step` of 1. This is far too high for the coordinates system that we use.

By calculating the coordinates fields `step` using the zoom level, we can ensure that the steps are nicely paced and useful on every zoom level.

Because the `defaultZoom` could be empty, we use the default fallback value of `8`. This default fallback value is also used by the preview component.